+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.1.14
+
+* Semantics of `Cleanup` have been changed. It can now be called multiple times
+  to register multiple cleanup functions. The cleanup functions are called in
+  the reverse order in which they were registered. Cleanup functions can no
+  longer return errors.
+
 ## v0.1.13
 
 * Added `PanicMiddleware` to `http` package.